requirements.txt
docxtpl==0.11.3
Jinja2==2.11.2
lxml==4.6.2
MarkupSafe==1.1.1
python-docx==0.8.10
pywin32==300
six==1.15.0
